+++ Samples/macOS-Swift/macOS-Swift/CppWrapper.h
@@ -1,7 +1,0 @@
-#import <Foundation/Foundation.h>
-
-@interface CppWrapper : NSObject
-- (void)throwCPPException;
-- (void)rethrowNoActiveCPPException;
-- (void)throwNSRangeException;
-@end